pyproject.toml
[project]
name = "10k-insight-agent"
version = "2.0.0"
description = "FastAPI + LangGraph microservice for analyzing SEC 10-K filings and matching to product catalog"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "langgraph>=0.2.50",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-groq>=0.2.0",
    "chromadb>=0.5.0",
    "beautifulsoup4>=4.12.0",
    "aiohttp>=3.10.0",
    "aiofiles>=24.1.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "pyyaml>=6.0.0",
    "streamlit>=1.39.0",
    "tiktoken>=0.7.0",
    "lxml>=5.3.0",
    "groq>=0.11.0",
    "sentence-transformers>=3.0.0",
    "torch>=2.0.0",
    "cohere>=5.9.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.13.0",
    "plotly>=5.18.0",
    "altair>=5.2.0",
    "pandas>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "httpx>=0.27.0",
]
